+++ pkgs/ffigen/test/header_parser_tests/record_use_test.dart
@@ -19,7 +19,7 @@ void main() {
         input: Input(entryPoints: [Uri.file(headerFile)]),
         visitors: [
           Visitor(
-            visitFunc: (node) {
+            func: (node) {
               node.isIncluded = true;
               node.recordUse = true;
               if (node.name == 'sum') {