+++ pkgs/ffigen/test/large_integration_tests/large_objc_test.dart
@@ -42,7 +42,7 @@ void main() {
     bool randInclude(String kind, DeclNode declaration, [String? member]) =>
         fnvHash32('$seed.$kind.${declaration.usr}.$member') <
         ((1 << 32) * inclusionRatio);
-    bool shouldIncludeNode(
+    bool stableRandomInclude(
       String kind,
       DeclNode declaration, [
       Set<String> forceIncludes = const {},
@@ -100,32 +100,34 @@ void main() {
       ),
       visitors: [
         Visitor(
-          visitFunc: (node) =>
-              node.isIncluded = shouldIncludeNode('functionDecl', node),
-          visitStruct: (node) =>
-              node.isIncluded = shouldIncludeNode('structDecl', node),
-          visitUnion: (node) =>
-              node.isIncluded = shouldIncludeNode('unionDecl', node),
-          visitEnum: (node) =>
-              node.isIncluded = shouldIncludeNode('enums', node),
-          visitUnnamedEnumConstant: (node) =>
-              node.isIncluded = shouldIncludeNode('unnamedEnumConstants', node),
-          visitGlobal: (node) =>
-              node.isIncluded = shouldIncludeNode('globals', node),
-          visitTypealias: (node) =>
-              node.isIncluded = shouldIncludeNode('typedefs', node),
-          visitMacro: (node) => node.isIncluded = false,
-          visitConstant: (node) => node.isIncluded = false,
-          visitObjCInterface: (node) =>
-              node.isIncluded = shouldIncludeNode('objcInterfaces', node),
-          visitObjCProtocol: (node) => node.isIncluded = shouldIncludeNode(
+          func: (node) =>
+              node.isIncluded = stableRandomInclude('functionDecl', node),
+          struct: (node) =>
+              node.isIncluded = stableRandomInclude('structDecl', node),
+          union: (node) =>
+              node.isIncluded = stableRandomInclude('unionDecl', node),
+          enumClass: (node) =>
+              node.isIncluded = stableRandomInclude('enums', node),
+          unnamedEnumConstant: (node) => node.isIncluded = stableRandomInclude(
+            'unnamedEnumConstants',
+            node,
+          ),
+          global: (node) =>
+              node.isIncluded = stableRandomInclude('globals', node),
+          typealias: (node) =>
+              node.isIncluded = stableRandomInclude('typedefs', node),
+          macroConstant: (node) => node.isIncluded = false,
+          constant: (node) => node.isIncluded = false,
+          objCInterface: (node) =>
+              node.isIncluded = stableRandomInclude('objcInterfaces', node),
+          objCProtocol: (node) => node.isIncluded = stableRandomInclude(
             'objcProtocols',
             node,
             forceIncludedProtocols,
           ),
-          visitObjCCategory: (node) =>
-              node.isIncluded = shouldIncludeNode('objcCategories', node),
-          visitObjCMethod: (node) {
+          objCCategory: (node) =>
+              node.isIncluded = stableRandomInclude('objcCategories', node),
+          objCMethod: (node) {
             final kind = switch (node.parent) {
               ObjCInterface() => 'objcInterfaces',
               ObjCProtocol() => 'objcProtocols',